lambda-v2/hospital_finder/requirements.txt
openai==0.28
requests
python-dotenv
